fix(train): temporarily disable logits and collate batch correctly
Unsloth does no longer output the logits and it required setting an
environment variable, which is stupid in itself, but then it doesn't
even work because it just breaks down as their weird "compiled" Python
code is littered with junk and has errors.
Great idea to define code in a string, where no static checker can check
anything...
All of the unsloth code gets more and more unhinged. I might look into
just using their kernels and ditching the rest, it's kind of a joke.