feat(trainer): handle console output with rich
Makes it easy to disable it on other processes, such that only the main
process shows the prints.
Also changed progress bars to use rich instead of tqdm. Same idea, that
there is no conflict going on, even though tqdm handles that quite well,
it just is nicer if it can be fully disabled for the other proceses.