feat(trainer): add spinner for tasks without progress bar
For example shows when it is saving the model, which takes some time but
does not show a progress bar.

This required a bit of a change, since rich does not allow to show
a spinner when a progress bar is already shown. To achieve this,
a custom Live element needs to be created, to which the status is added
next to the progress bar, as there can only ever be a single Live
element.