feat(dataset): provide option to resize images
Some models use the full resolution images, such as Qwen2-VL, or base
their patching on the original image, therefore changing the size of the
images can have a noticeable impact.

The resizing is aspect ratio preserving, where the larger side is
resized to the specified size while the smaller side is resized
according to the aspect ratio.
In addition to specifying the size, it is possible to also define when
the resizing should be performed. In particular, the options
"when-smaller" and "when-larger" can be used to ensure a minimum or
maximum size, respectively.