chore: configure basedpyright to automatically look in the venv
basedpyright may not look into the venv for packages, which means that
you need to have the venv active for it to find all imported packages.
This is generally not an issue, since it is run with uv anyway, but for
some editors this makes it much easier to deal with the venv.