fix(chat): unpack the singular batch after tokenising to identify the assistant
For some reason, now it is always batched, whereas before, it would
automatically unpack it, if a single conversation was given.
Not sure if that will stay like that or if it was a bug that was
introduced.
Regardless, this also fixes some annoying type issues with having
a union as output that depends on the inputs. Also, if it breaks again,
only one place needs to be modified.