chore(dependencies): define dependencies for aarch64
aarch64 + CUDA needs to compile most packages from source as there are
no distributions available for that combination. uv handles that quite
well, but there are some issues that need to be taken care of by the
user.

NOTE: the extension of bitsandbytes needs to be compiled by hand until
the PR is merged that will take care of it.